Create more polished home. Implement fetch user data from database. Changed transactions from example to related by user.

+++ backend/src/main/java/com/github/sdurell/budgeting_app/model/UserEntity.java
@@ -24,7 +24,7 @@ public class UserEntity {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String username;
     private String password;
 
@@ -33,7 +33,4 @@ public class UserEntity {
             joinColumns = @JoinColumn(name = "user_id", referencedColumnName = "id"),
             inverseJoinColumns = @JoinColumn(name = "role_id", referencedColumnName = "id"))
     private List<Role> roles = new ArrayList<>();
-
-    // @ManyToOne(fetch = FetchType.EAGER, cascade = CascadeType.ALL)
-    // @JoinTable(name = "user_transactions")
 }
